pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aicli"
version = "0.1.0"
description = "Universal agentic CLI for large language models"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27",
    "pyyaml>=6.0",
    "rich>=13.0",
    "click>=8.1",
]

[project.optional-dependencies]
gemini = ["google-generativeai>=0.8"]
anthropic = ["anthropic>=0.40"]
openai = ["openai>=1.50"]

[project.scripts]
aicli = "aicli.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
